create component for listing files
